[Bronze I] Title: 평균, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/1546. 평균/평균.cc
@@ -20,6 +20,6 @@ int main(void)
 		a[i] = (a[i] / max) * 100;
 		sum += a[i];
 	}
-	cout << sum / (float)n;
+	cout << sum / n;
 	return 0;
 }
